requirements.txt
alembic==1.7.4
bcrypt==3.2.0
bidict==0.21.4
blinker==1.4
cffi==1.15.0
click==7.1.2
coverage==6.1
cryptography==35.0.0
dnspython==2.1.0
Flask==1.1.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.4.0
Flask-User==1.0.1.5
Flask-WTF==0.14.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
importlib-metadata==4.8.1
importlib-resources==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.5
MarkupSafe==1.1.1
passlib==1.7.4
py==1.10.0
pycparser==2.20
pycryptodome==3.11.0
pymongo==3.12.1
pytest==3.0.5
pytest-cov==2.4.0
python-engineio==4.3.0
python-socketio==5.4.1
six==1.16.0
SQLAlchemy==1.4.26
Werkzeug==1.0.1
wsproto==1.0.0
WTForms==2.3.3
zipp==3.6.0
